Day 2: Exercises 3 through 6
Making good progress so far. The book is very beginner level-isa so it
makes for light reading when I’m off javascript. If I keep up with 3
exercises a day, I should be done in 17 days which is perfect. I may
try to go for four but I want my early stages of python to be about
fun. I still have lots of JS to cover.