Day 3
Pushed through 4 exercises. The target is to do 10 over the weekend. 5
a day on Saturday and Sunday each. That will bring me to 23 of the
exercises done within a week which means I may just finish this book in
two weeks. On a technical note, it’s fun to write python. It’s nice to
have something to compare Javascript to.